requirements.txt
torch
numpy
Pillow
streamlit